+++ sdk/textanalytics/azure-ai-textanalytics/src/samples/java/com/azure/ai/textanalytics/TextAnalyticsAsyncClientJavaDocCodeSnippets.java
@@ -19,8 +19,6 @@ import com.azure.ai.textanalytics.models.HealthcareEntity;
 import com.azure.ai.textanalytics.models.PiiEntityCollection;
 import com.azure.ai.textanalytics.models.PiiEntityDomain;
 import com.azure.ai.textanalytics.models.RecognizeEntitiesAction;
-import com.azure.ai.textanalytics.models.RecognizeEntitiesOptions;
-import com.azure.ai.textanalytics.models.RecognizeLinkedEntitiesOptions;
 import com.azure.ai.textanalytics.models.RecognizePiiEntitiesOptions;
 import com.azure.ai.textanalytics.models.SentenceSentiment;
 import com.azure.ai.textanalytics.models.TargetSentiment;
@@ -253,39 +251,6 @@ public class TextAnalyticsAsyncClientJavaDocCodeSnippets {
         // END: com.azure.ai.textanalytics.TextAnalyticsAsyncClient.recognizeCategorizedEntitiesBatch#Iterable-TextAnalyticsRequestOptions
     }
 
-    /**
-     * Code snippet for {@link TextAnalyticsAsyncClient#recognizeEntitiesBatchWithResponse(Iterable, RecognizeEntitiesOptions)}
-     */
-    public void recognizeBatchEntitiesMaxOverloadWithRecognizeEntitiesOptions() {
-        // BEGIN: com.azure.ai.textanalytics.TextAnalyticsAsyncClient.recognizeCategorizedEntitiesBatch#Iterable-RecognizeEntitiesOptions
-        List<TextDocumentInput> textDocumentInputs1 = Arrays.asList(
-            new TextDocumentInput("0", "I had a wonderful trip to Seattle last week.").setLanguage("en"),
-            new TextDocumentInput("1", "I work at Microsoft.").setLanguage("en"));
-
-        RecognizeEntitiesOptions requestOptions = new RecognizeEntitiesOptions().setIncludeStatistics(true);
-
-        textAnalyticsAsyncClient.recognizeEntitiesBatchWithResponse(textDocumentInputs1, requestOptions)
-            .subscribe(response -> {
-                // Response's status code
-                System.out.printf("Status code of request response: %d%n", response.getStatusCode());
-                RecognizeEntitiesResultCollection resultCollection = response.getValue();
-
-                // Batch statistics
-                TextDocumentBatchStatistics batchStatistics = resultCollection.getStatistics();
-                System.out.printf("Batch statistics, transaction count: %s, valid document count: %s.%n",
-                    batchStatistics.getTransactionCount(), batchStatistics.getValidDocumentCount());
-
-                resultCollection.forEach(
-                    recognizeEntitiesResult -> recognizeEntitiesResult.getEntities().forEach(
-                        entity -> System.out.printf(
-                            "Recognized categorized entity: %s, category: %s, confidence score: %f.%n",
-                            entity.getText(),
-                            entity.getCategory(),
-                            entity.getConfidenceScore())));
-            });
-        // END: com.azure.ai.textanalytics.TextAnalyticsAsyncClient.recognizeCategorizedEntitiesBatch#Iterable-RecognizeEntitiesOptions
-    }
-
     // Personally Identifiable Information Entity
 
     /**
@@ -514,43 +479,6 @@ public class TextAnalyticsAsyncClientJavaDocCodeSnippets {
                     }));
             });
         // END: com.azure.ai.textanalytics.TextAnalyticsAsyncClient.recognizeLinkedEntitiesBatch#Iterable-TextAnalyticsRequestOptions
-    }
-
-    /**
-     * Code snippet for {@link TextAnalyticsAsyncClient#recognizeLinkedEntitiesBatchWithResponse(Iterable, RecognizeLinkedEntitiesOptions)}
-     */
-    public void recognizeBatchLinkedEntitiesMaxOverloadWithRecognizeLinkedEntitiesOptions() {
-        // BEGIN: com.azure.ai.textanalytics.TextAnalyticsAsyncClient.recognizeLinkedEntitiesBatch#Iterable-RecognizeLinkedEntitiesOptions
-        List<TextDocumentInput> textDocumentInputs1 = Arrays.asList(
-            new TextDocumentInput("0", "Old Faithful is a geyser at Yellowstone Park.").setLanguage("en"),
-            new TextDocumentInput("1", "Mount Shasta has lenticular clouds.").setLanguage("en"));
-
-        RecognizeLinkedEntitiesOptions requestOptions = new RecognizeLinkedEntitiesOptions().setIncludeStatistics(true);
-
-        textAnalyticsAsyncClient.recognizeLinkedEntitiesBatchWithResponse(textDocumentInputs1, requestOptions)
-            .subscribe(response -> {
-                // Response's status code
-                System.out.printf("Status code of request response: %d%n", response.getStatusCode());
-                RecognizeLinkedEntitiesResultCollection resultCollection = response.getValue();
-
-                // Batch statistics
-                TextDocumentBatchStatistics batchStatistics = resultCollection.getStatistics();
-                System.out.printf("Batch statistics, transaction count: %s, valid document count: %s.%n",
-                    batchStatistics.getTransactionCount(), batchStatistics.getValidDocumentCount());
-
-                resultCollection.forEach(
-                    recognizeLinkedEntitiesResult -> recognizeLinkedEntitiesResult.getEntities().forEach(
-                        linkedEntity -> {
-                            System.out.println("Linked Entities:");
-                            System.out.printf("Name: %s, entity ID in data source: %s, URL: %s, data source: %s.%n",
-                                linkedEntity.getName(), linkedEntity.getDataSourceEntityId(), linkedEntity.getUrl(),
-                                linkedEntity.getDataSource());
-                            linkedEntity.getMatches().forEach(entityMatch -> System.out.printf(
-                                "Matched entity: %s, confidence score: %.2f.%n",
-                                entityMatch.getText(), entityMatch.getConfidenceScore()));
-                        }));
-            });
-        // END: com.azure.ai.textanalytics.TextAnalyticsAsyncClient.recognizeLinkedEntitiesBatch#Iterable-RecognizeLinkedEntitiesOptions
     }
 
     // Key Phrases